aws_ecr_repository: Avoid second API request when creating
See issue #3849

In some cases, AWS can answer with an error 400
RepositoryNotFoundException on te second DescribeRepositories request
right after the repository was successfully created. The resource end up
with repository_url not there (can cause errors if it is used elsewhere)
and the tfstate says that the ECR does not exists while it was in fact
created. We do not need to perform this request, and can avoid this
error.